Implement §5.2.1.1 Trigraph sequences
I'm thinking of the preprocessor as a pipelined filter, with each
filter stage applying a transformation independently of the others.
So trigraphs are replaced in one stage, continued lines are spliced
in another stage, preprocessing tokens are generated in yet another
stage, and preprocessing directives are executed in a final stage
that produces tokens for the parser.

Each filter stage implements the StreamFilter interface, exposing
`get()` and `putback()` methods. To compose the filter stages into
a pipeline, the StreamFilter constructor takes an input Stream
reference. You could compose these arbitrarily at run-time with
something like:

    auto pipe = new StreamFilter(
                    new StreamFilter(
                        new StreamFilter(
                            new StreamFilter(
                                ... ad nauseum;

but for the C preprocessor, we know at compile-time how we want to
compose the pipeline, so there's no need to use dynamic memory
allocation. Instead, I declare reference members for each stage of
the pipeline and compose them in the member initializer list.

There's probably a better way.