Implement §6.4.2.1 Identifiers
I'm introducing a LexicalAnalyzer class to tokenize the input stream,
and a PPToken struct to represent the preprocessing tokens given in
§6.4.

Universal character names in identifiers are not currently supported,
and there are no "other implementation-defined characters".

Since we're introducing another test file, it's time to extract the
Catch main routine to its own translation unit.

-Wno-c++98-compat-pedantic is necessary to avoid a warning about
trailing commas in enum definitions.